// server/types/tmdb.ts
export interface TmdbMovieDto {
  adult: boolean
  backdrop_path: string | null
  genre_ids: number[]
  id: number
  title: string
  original_language: string
  original_title: string
  overview: string
  popularity: number
  poster_path: string | null
  release_date: string
  video: boolean
  vote_average: number
  vote_count: number
}

export interface TmdbPaginatedResponse<T> {
  page: number
  results: T[]
  total_pages: number
  total_results: number
}